Feature/recommend queue 推荐算法执行器；推荐数据接口

+++ src/main/java/com/chenlinghong/graduation/repository/domain/RecommendRankingGoods.java
@@ -30,5 +30,5 @@ public class RecommendRankingGoods extends BaseDomain {
     /**
      * 排行榜
      */
-    private Double ranking;
+    private Integer ranking;
 }
